+++ src/components/Review/ViewReviewModal.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import Modal from '../../components/Modal/Modal';
+import Modal from '../Modal/Modal';
 import styled from 'styled-components';
-import infoImg from './info-image.jpg';
-import Star from './Star';
+import infoImg from '../../pages/restaurantInfo/info-image.jpg';
+import Star from '../Star/Star';
 
 const ViewReviewModal = ({ closeModal }) => {
   return (
