+++ src/components/Star/Star.tsx
